Fix race condition
Make sure to call new connection delegate first before wiring
up the connection adapter, because the new connection
delegate is going to be setting the data received and
connection broken delegates.

If we were to wire up the connection adapter first, that kicks
off a worker thread in SystemAbstractions::NetworkConnection,
which will start to read the underlying socket, and if data
happens to come in right away, it will never reach the data
received delegate, because that hasn't been set yet.