Server: extra protection when calling data received and broken delegates
If the delegates are replaced during the delegate calls themselves,
ensure the delegate states (e.g. lambda expression captures) aren't
prematurely destroyed, by making a copy of the delegate and calling
through the copy.

We also don't need to hold the mutex when actually calling the
delegate, now that we're calling through a local copy.